requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
djangorestframework==3.11.0
html5lib==1.0.1
idna==2.9
Pillow==7.0.0
PyJWT==1.7.1
PyPDF2==1.26.0
pytz==2019.3
reportlab==3.5.34
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
twilio==6.35.5
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.0
xhtml2pdf==0.2.4
